fix: cambiar FileInode de uint64 a int64 para compatibilidad con SQLite

+++ pkg/filesystem/scanner.go
@@ -14,7 +14,7 @@ import (
 type FileEntry struct {
 	Path          string   // Absolute path to the file
 	Size          int64    // File size in bytes
-	Inode         uint64   // Inode number (for hardlink detection)
+	Inode         int64    // Inode number (for hardlink detection)
 	ModTime       int64    // Modification time (Unix timestamp)
 	IsHardlink    bool     // Whether this file has hardlinks
 	HardlinkPaths []string // All paths that are hardlinks to this file
@@ -38,7 +38,7 @@ type Scanner struct {
 	logger  *zap.Logger
 
 	// Internal state
-	inodeMap map[uint64][]*FileEntry // Map of inode -> list of paths (for hardlink detection)
+	inodeMap map[int64][]*FileEntry // Map of inode -> list of paths (for hardlink detection)
 	mu       sync.RWMutex
 }
 
@@ -55,7 +55,7 @@ func NewScanner(options ScanOptions, logger *zap.Logger) *Scanner {
 	return &Scanner{
 		options:  options,
 		logger:   logger,
-		inodeMap: make(map[uint64][]*FileEntry),
+		inodeMap: make(map[int64][]*FileEntry),
 	}
 }
 
@@ -146,7 +146,7 @@ func (s *Scanner) scanPath(rootPath string, result map[string]*FileEntry) error 
 		entry := &FileEntry{
 			Path:        path,
 			Size:        info.Size(),
-			Inode:       stat.Ino,
+			Inode:       int64(stat.Ino),
 			ModTime:     info.ModTime().Unix(),
 			MediaType:   s.inferMediaType(path),
 			PrimaryPath: path, // Will be updated in second pass
@@ -157,7 +157,7 @@ func (s *Scanner) scanPath(rootPath string, result map[string]*FileEntry) error 
 
 		// Add to inode map for hardlink detection
 		s.mu.Lock()
-		s.inodeMap[stat.Ino] = append(s.inodeMap[stat.Ino], entry)
+		s.inodeMap[int64(stat.Ino)] = append(s.inodeMap[int64(stat.Ino)], entry)
 		s.mu.Unlock()
 
 		return nil
@@ -234,7 +234,7 @@ func (s *Scanner) detectHardlinks(entries map[string]*FileEntry) {
 		}
 
 		s.logger.Debug("Detected hardlink group",
-			zap.Uint64("inode", inode),
+			zap.Int64("inode", inode),
 			zap.Int("count", len(files)),
 			zap.String("primary", primaryPath),
 			zap.Strings("all_paths", allPaths),
